+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Dev
+
+### Fixed
+- Fixed Transmutation inventory not syncing correctly due to code being called on the client
+
 ## 1.0.8
 
 ### Fixed